Added working login function and provided some basic styling

+++ J2EE/src/main/java/nu/educom/rvt/models/Role.java
@@ -8,7 +8,7 @@ public class Role {
 
 	@Id
 	@GeneratedValue
-	private long id;
+	private int id;
 	private String name;
 	
 	public Role() {
@@ -20,13 +20,15 @@ public class Role {
 		this.name = name;
 	}
 	
-	public long getId() {
+	
+	public int getId() {
 		return id;
 	}
-	
-	public void setId(long id) {
+
+	public void setId(int id) {
 		this.id = id;
 	}
+
 	public String getName() {
 		return name;
 	}
